requirements.txt
streamlit==1.44.1
fpdf2==2.8.2
polars==1.26.0
